+++ packages/eslint-config-react/index.js
@@ -5,11 +5,9 @@ module.exports = {
 		'./rules/jsx-a11y',
 	].map(require.resolve),
 	overrides: {
-		{
-			"files": ["**/*.tsx"],
-			"rules": {
-				"react/prop-types": "off"
-			}
-		}
-	}
+		files: ['**/*.tsx'],
+		rules: {
+			'react/prop-types': 'off',
+		},
+	},
 };